Use `ENDPOINT_CALLBACK` for extra_endpoints.
