+++ apollo-router/src/axum_factory/axum_http_server_factory.rs
@@ -4,7 +4,6 @@ use std::sync::atomic::AtomicBool;
 use std::sync::atomic::AtomicU64;
 use std::sync::atomic::Ordering;
 use std::sync::Arc;
-use std::sync::OnceLock;
 use std::time::Instant;
 
 use axum::extract::Extension;
@@ -43,6 +42,7 @@ use super::listeners::ListenersAndRouters;
 use super::utils::decompress_request_body;
 use super::utils::PropagatingMakeSpan;
 use super::ListenAddrAndRouter;
+use super::ENDPOINT_CALLBACK;
 use crate::axum_factory::compression::Compressor;
 use crate::axum_factory::listeners::get_extra_listeners;
 use crate::axum_factory::listeners::serve_router_on_listen_addr;
@@ -451,19 +451,6 @@ where
 
     let listener = configuration.supergraph.listen.clone();
     Ok(ListenAddrAndRouter(listener, route))
-}
-
-static ENDPOINT_CALLBACK: OnceLock<Arc<dyn Fn(Router) -> Router + Send + Sync>> = OnceLock::new();
-
-/// Set a callback that may wrap or mutate `axum::Router` as they are added to the main router.
-/// Although part of the public API, this is not intended for use by end users, and may change at any time.
-#[doc(hidden)]
-pub fn unsupported_set_axum_router_callback(
-    callback: impl Fn(Router) -> Router + Send + Sync + 'static,
-) -> Result<(), &'static str> {
-    ENDPOINT_CALLBACK
-        .set(Arc::new(callback))
-        .map_err(|_| "endpoint decorator was already set")
 }
 
 async fn metrics_handler<B>(request: Request<B>, next: Next<B>) -> Response {
